third_party/WebKit/LayoutTests/fast/dom/dataset-gc-expected.txt

This tests that custom properties on element.dataset persist GC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS d.getAttribute('data-custom-property') is '1'
PASS d.dataset.customProperty is '1'
PASS d.dataset.customProperty is '1'
PASS successfullyParsed is true

TEST COMPLETE
